fix: 验证码错误提示

+++ apps/authentication/services.py
@@ -32,21 +32,25 @@ def exchange_code(code: str, redirect_uri):
 
 def check_chaptcha(captcha_id, captcha_code):
     if captcha_id is None:
-        raise ValidationError(message="验证码错误")
+        raise ValidationError(message="Captcha code error, please refresh the page.")
     if captcha_code is None:
-        raise ValidationError(message="验证码错误")
-    expiration = CaptchaStore.objects.filter(id=captcha_id).first().expiration
-    expiration = expiration.astimezone(pytz.timezone("Asia/Shanghai"))
-    response = CaptchaStore.objects.filter(id=captcha_id).first().response
-    image_code = CaptchaStore.objects.filter(id=captcha_id).first()
-    five_minute_ago = datetime.now() - timedelta(hours=0, minutes=5, seconds=0)
-    five_minute_ago = five_minute_ago.replace(tzinfo=pytz.timezone("Asia/Shanghai"))
-    if image_code and five_minute_ago > expiration:
-        image_code.delete()
-        raise ValidationError(message="验证码过期")
-    else:
-        if image_code and response.lower() == captcha_code.lower():
+        raise ValidationError(message="Captcha code error, please refresh the page.")
+    code_obj=CaptchaStore.objects.filter(id=captcha_id)
+    if code_obj.exists():
+        expiration = code_obj.first().expiration
+        expiration = expiration.astimezone(pytz.timezone("Asia/Shanghai"))
+        response = CaptchaStore.objects.filter(id=captcha_id).first().response
+        image_code = CaptchaStore.objects.filter(id=captcha_id).first()
+        five_minute_ago = datetime.now() - timedelta(hours=0, minutes=5, seconds=0)
+        five_minute_ago = five_minute_ago.replace(tzinfo=pytz.timezone("Asia/Shanghai"))
+        if image_code and five_minute_ago > expiration:
             image_code.delete()
+            raise ValidationError(message="Captcha code error, please refresh the page.")
         else:
-            image_code.delete()
-            raise ValidationError(message="验证码错误")
+            if image_code and response.lower() == captcha_code.lower():
+                image_code.delete()
+            else:
+                image_code.delete()
+                raise ValidationError(message="Captcha code error, please refresh the page.")
+    else:
+        raise ValidationError(message="Captcha code error, please refresh the page.")
